feat(ironfish): Yeet the message routers (#1209)
* feat(ironfish): Yeet the message routers

* Fix lint
