--- ironfish/src/network/messageRouters/rpc.test.ts
@@ -1,148 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-jest.mock('./rpcId')
-import { mocked } from 'ts-jest/utils'
-import { CannotSatisfyRequest } from '../messages/cannotSatisfyRequest'
-import { GetBlockHashesRequest, GetBlockHashesResponse } from '../messages/getBlockHashes'
-import { NetworkMessageType } from '../messages/networkMessage'
-import { NetworkError } from '../peers/connections/errors'
-import { PeerManager } from '../peers/peerManager'
-import { getConnectedPeer, mockHostsStore, mockLocalPeer } from '../testUtilities'
-import { CannotSatisfyRequestError, RequestTimeoutError, RpcRouter } from './rpc'
-import { nextRpcId, rpcTimeoutMillis } from './rpcId'
-
-describe('RPC Router', () => {
-  beforeEach(() => {
-    jest.resetAllMocks()
-    jest.useFakeTimers()
-  })
-
-  beforeEach(() => {
-    mocked(nextRpcId).mockReturnValue(91)
-    mocked(rpcTimeoutMillis).mockReturnValue(10)
-  })
-
-  it('Registers an RPC Handler', () => {
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-    const router = new RpcRouter(peers)
-    const handler = jest.fn()
-    router.register(NetworkMessageType.PeerList, handler)
-    expect(router['handlers'].size).toBe(1)
-    expect(router['handlers'].get(NetworkMessageType.PeerList)).toBe(handler)
-  })
-
-  it('should time out RPC requests', async () => {
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-
-    const { peer } = getConnectedPeer(peers)
-    const sendMock = jest.spyOn(peer, 'send')
-    const peerCloseMock = jest.spyOn(peer, 'close')
-
-    const router = new RpcRouter(peers)
-    const handlerMock = jest.fn()
-    router.register(NetworkMessageType.GetBlockHashesRequest, handlerMock)
-    expect(router['requests'].size).toBe(0)
-
-    const promise = router.requestFrom(peer, new GetBlockHashesRequest(0, 0, 0))
-
-    expect(router['requests'].size).toBe(1)
-    jest.runOnlyPendingTimers()
-
-    expect(router['requests'].size).toBe(0)
-    expect(sendMock).toHaveBeenCalledTimes(1)
-    expect(peerCloseMock).toHaveBeenCalled()
-    await expect(promise).toRejectErrorInstance(RequestTimeoutError)
-  })
-
-  it('should reject requests when connection disconnects', async () => {
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-
-    const { peer, connection } = getConnectedPeer(peers)
-    const peerCloseMock = jest.spyOn(peer, 'close')
-    const sendMock = jest.spyOn(peer, 'send')
-
-    const router = new RpcRouter(peers)
-    router.register(NetworkMessageType.GetBlockHashesRequest, jest.fn())
-    expect(router['requests'].size).toBe(0)
-
-    const subscribers = peer.onStateChanged.subscribers
-
-    const promise = router.requestFrom(peer, new GetBlockHashesRequest(0, 0, 0))
-
-    expect(router['requests'].size).toBe(1)
-    expect(peer.onStateChanged.subscribers).toBeGreaterThan(subscribers)
-    connection.close()
-
-    expect(peer.onStateChanged.subscribers).toBeLessThanOrEqual(subscribers)
-    expect(router['requests'].size).toBe(0)
-    expect(sendMock).toHaveBeenCalledTimes(1)
-    expect(peerCloseMock).not.toHaveBeenCalled()
-    await expect(promise).toRejectErrorInstance(NetworkError)
-  })
-
-  it('should increment and decrement pendingRPC', async () => {
-    mocked(nextRpcId).mockReturnValue(91)
-
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-    jest.spyOn(peers, 'sendTo')
-    const { peer } = getConnectedPeer(peers, 'peer')
-
-    const router = new RpcRouter(peers)
-    router.register(NetworkMessageType.GetBlockHashesRequest, jest.fn())
-
-    const rpcId = 91
-    void router.requestFrom(peer, new GetBlockHashesRequest(0, 0, rpcId))
-    expect(peer.pendingRPC).toBe(1)
-
-    const response = new GetBlockHashesResponse(['chili'], rpcId)
-    await router.handle(peer, response)
-    expect(peer.pendingRPC).toBe(0)
-  })
-
-  it('Handles a response as a resolved request promise', async () => {
-    mocked(nextRpcId).mockReturnValue(91)
-    mocked(rpcTimeoutMillis).mockReturnValue(1000)
-
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-
-    const router = new RpcRouter(peers)
-    router.register(NetworkMessageType.GetBlockHashesRequest, jest.fn())
-
-    const { peer: peer1 } = getConnectedPeer(peers)
-    const { peer: peer2 } = getConnectedPeer(peers)
-
-    const rpcId = 91
-    const promise = router.requestFrom(peer1, new GetBlockHashesRequest(0, 0, rpcId))
-    const response = new GetBlockHashesResponse(['chili'], rpcId)
-
-    await router.handle(peer2, response)
-    await expect(promise).resolves.toMatchObject({
-      message: response,
-    })
-
-    expect(router['requests'].size).toBe(0)
-  })
-
-  it('Catches a cannotSatisfy error and returns the appropriate type', async () => {
-    mocked(nextRpcId).mockReturnValue(18)
-
-    const peers = new PeerManager(mockLocalPeer(), mockHostsStore())
-
-    const handlerMock = jest.fn(() => {
-      throw new CannotSatisfyRequestError('Bad request')
-    })
-    const router = new RpcRouter(peers)
-    router.register(NetworkMessageType.GetBlockHashesRequest, handlerMock)
-
-    const { peer } = getConnectedPeer(peers)
-    const sendMock = jest.spyOn(peer, 'send')
-    const rpcId = 18
-    await router.handle(peer, new GetBlockHashesRequest(0, 0, rpcId))
-
-    expect(router['requests'].size).toBe(0)
-    expect(sendMock).toBeCalledTimes(1)
-    expect(sendMock).toHaveBeenCalledWith(new CannotSatisfyRequest(rpcId))
-  })
-})
